pyproject.toml
[project]
name = "clinisight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "groq>=1.0.0",
    "lxml>=6.0.2",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
